+++ src/constants.ts
@@ -21,8 +21,6 @@ export const COMMON_HEADERS = {
   Connection: "keep-alive",
 };
 
-export const MAX_REDIRECTS = 5;
-
 // Regex
 export const REGEX_TITLE_TAG = /<title[^>]*>([^<]+)<\/title>/i;
 export const REGEX_SIMPLE_HTML_TITLE_FALLBACK = /<html>([^<]+)<\/html>/;